Corrected web server issue, updated files, changed command mode, groundwork for attachmend mode
