+++ motionTask.c
@@ -25,7 +25,6 @@ void motionInit()
 void motionForward()
 {
     thermoSensorFlag = true;
-    motion_init();
     motion_servo_set_pulse_width(MOTION_WHEEL_RIGHT, 1100);
     motion_servo_set_pulse_width(MOTION_WHEEL_LEFT, 4800);
     motion_servo_set_pulse_width(MOTION_SERVO_CENTER, 1400);
@@ -56,6 +55,19 @@ void motionSpinLeft()
     thermoSensorFlag = false;
     motion_servo_set_pulse_width(MOTION_WHEEL_RIGHT, 1100);
     motion_servo_set_pulse_width(MOTION_WHEEL_LEFT, 1100);
+
+    motion_servo_start(MOTION_WHEEL_RIGHT);
+    motion_servo_start(MOTION_WHEEL_LEFT);
+}
+
+/**
+ * Spin Chico to the left slowly
+ */
+void motionSpinLeftSlow()
+{
+    thermoSensorFlag = false;
+    motion_servo_set_pulse_width(MOTION_WHEEL_RIGHT, 2550);
+    motion_servo_set_pulse_width(MOTION_WHEEL_LEFT, 2550);
 
     motion_servo_start(MOTION_WHEEL_RIGHT);
     motion_servo_start(MOTION_WHEEL_LEFT);
