+++ temperatureHandler.h
@@ -2,6 +2,11 @@
 #ifndef TEMPERATUREHANDLER_H_
 #define TEMPERATUREHANDLER_H_
 
+int getLeft3AvgTemperatures();
+int getRight3AvgTemperatures();
+int getCenter4AvgTemperatures();
+int getAmbientTemperature();
+
 void setupTemperature();
 void updateTemperatures(int *ambientTemperature, int *leftAverageTemperature, int *rightAverageTemperature);
